Detail that git context ignores .dockerignore
The issue is detailed here https://github.com/docker/build-push-action/issues/182

My team lost hours because of .dockerignore is ignored unless `context: .` and now must update many projects. If this mention of .dockerignore had been in place we likely would have been clued in that git context is not what we want at all and we had to be in path context.

Hopefully this helps other developers.